refactor: apply hook folder rename
